+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 - Remove `experimental` tag for  `ml_client.jobs.validate`.
+- Workspace Connection has new read-only subclass: AzureBlobStoreWorkspaceConnectionSchema.
+- Workspace Connection supports 2 new types under main class: adls_gen_2 and azure_one_lake.
+- Workspace connection supports reading data connections.
 
 ### Bugs Fixed
 - Fix pipeline job `outputs` not load correctly when `component: <local-file>` exists in pipeline job yaml.